Show the issue with a test
