+++ src/Codec/Xlsx/Types/Common.hs
@@ -1,5 +1,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE TemplateHaskell #-}
+
 module Codec.Xlsx.Types.Common
   ( CellRef(..)
   , singleCellRef
@@ -21,6 +23,14 @@ module Codec.Xlsx.Types.Common
   , col2int
   , RowIndex
   , ColIndex
+  -- ** prisms
+  , _XlsxText
+  , _XlsxRichText
+  , _CellText
+  , _CellDouble
+  , _CellBool
+  , _CellRich
+  , _CellError
   ) where
 
 import GHC.Generics (Generic)
@@ -42,6 +52,7 @@ import Safe
 import Text.XML
 import Text.XML.Cursor
 
+import Control.Lens(makePrisms)
 import Codec.Xlsx.Parser.Internal
 import Codec.Xlsx.Types.RichText
 import Codec.Xlsx.Writer.Internal
@@ -146,6 +157,7 @@ data XlsxText = XlsxText Text
               | XlsxRichText [RichTextRun]
               deriving (Eq, Ord, Show, Generic)
 
+
 instance NFData XlsxText
 
 xlsxTextToCellValue :: XlsxText -> CellValue
@@ -171,6 +183,7 @@ data CellValue
   | CellRich [RichTextRun]
   | CellError ErrorType
   deriving (Eq, Ord, Show, Generic)
+
 
 instance NFData CellValue
 
@@ -391,3 +404,6 @@ instance ToAttrVal ErrorType where
 
 type RowIndex = Int
 type ColIndex = Int
+
+makePrisms ''XlsxText
+makePrisms ''CellValue
